pyproject.toml
[tool.poetry]

name = "memegen-flet"
version = "0.1"
description = "Memegen.link client built with Flet."
authors = ["Jace Browning <jacebrowning@gmail.com>"]

[tool.poetry.dependencies]

python = "~3.11"

flet = "~0.22"
minilog = "^2.1"
requests = "^2.28.1"
universal-startfile = "^0.2"

[tool.poetry.dev-dependencies]

black = "^23.7"
isort = "^5.12"

pyinstaller = "^6.6"
pillow = "^10.0"

[build-system]

requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
